requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.14.0
requests
requests-toolbelt==0.9.1
urllib3==1.25.10
webexteamssdk
Pillow
gunicorn
prettyprinter

